@@ -1,6 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-import TestUtils from 'react-addons-test-utils';
+import ReactTestUtils from 'react-dom/test-utils';
 import Goljus from './Goljus';
 
 const diff = require('jest-diff');
@@ -8,28 +7,28 @@ const diff = require('jest-diff');
 jest.useFakeTimers();
 
 it('renders without crashing', () => {
-    let goljus = TestUtils.renderIntoDocument(<Goljus />);
+    let goljus = ReactTestUtils.renderIntoDocument(<Goljus />);
 });
 
 it('has current board state', () => {
-    let goljus = TestUtils.renderIntoDocument(<Goljus />);
+    let goljus = ReactTestUtils.renderIntoDocument(<Goljus />);
     expect(goljus.state.board).toBeDefined();
 });
 
 it('has board state size that depends on props', () => {
-    let goljus = TestUtils.renderIntoDocument(<Goljus shape="5,5" />);
+    let goljus = ReactTestUtils.renderIntoDocument(<Goljus shape="5,5" />);
     expect(goljus.state.board.length).toEqual(5);
     expect(goljus.state.board[0].length).toEqual(5);
 });
 
 it('can have board state size that is not square', () => {
-    let goljus = TestUtils.renderIntoDocument(<Goljus shape="5,6" />);
+    let goljus = ReactTestUtils.renderIntoDocument(<Goljus shape="5,6" />);
     expect(goljus.state.board.length).toEqual(5);
     expect(goljus.state.board[0].length).toEqual(6);
 });
 
 it('has a board by default', () => {
-    let goljus = TestUtils.renderIntoDocument(<Goljus/>);
+    let goljus = ReactTestUtils.renderIntoDocument(<Goljus/>);
     expect(goljus.state.board.length).toEqual(30);
     expect(goljus.state.board[0].length).toEqual(30);
 });
@@ -78,7 +77,7 @@ expect.extend({
 });
 
 it('can give the board to a caller', () => {
-    let goljus = TestUtils.renderIntoDocument(<Goljus shape="2,3"/>);
+    let goljus = ReactTestUtils.renderIntoDocument(<Goljus shape="2,3"/>);
     expect(goljus.getBoard).toBeDefined();
     expect(goljus.getBoard()).toBeSimilarWithBoard(Goljus.createBoard(2, 3));
     expect(goljus.getBoard().height).toEqual(3);
@@ -86,14 +85,14 @@ it('can give the board to a caller', () => {
 });
 
 it('can update the board', () => {
-    let goljus = TestUtils.renderIntoDocument(<Goljus shape="5,5"/>);
+    let goljus = ReactTestUtils.renderIntoDocument(<Goljus shape="5,5"/>);
     expect(Goljus.updateBoard).toBeDefined();
     let oldBoard = goljus.getBoard();
     expect(Goljus.updateBoard(oldBoard)).toEqual(oldBoard);
 });
 
 it('kills cells with too few neighbors on update', () => {
-    let goljus = TestUtils.renderIntoDocument(<Goljus shape="5,5"/>);
+    let goljus = ReactTestUtils.renderIntoDocument(<Goljus shape="5,5"/>);
     expect(Goljus.updateBoard).not.toBe(undefined);
     let oldBoard = goljus.getBoard();
     oldBoard[2][2] = true;
@@ -103,7 +102,7 @@ it('kills cells with too few neighbors on update', () => {
 });
 
 it('kills cells with too many neighbors on update', () => {
-    let goljus = TestUtils.renderIntoDocument(<Goljus shape="5,5"/>);
+    let goljus = ReactTestUtils.renderIntoDocument(<Goljus shape="5,5"/>);
     expect(Goljus.updateBoard).not.toBe(undefined);
     let oldBoard = goljus.getBoard();
     oldBoard[0][1] = true;
@@ -168,7 +167,7 @@ it('has a cyclic form', () => {
 });
 
 it('has an advancing control', () => {
-    let goljus = TestUtils.renderIntoDocument(<Goljus shape="5,5"/>);
+    let goljus = ReactTestUtils.renderIntoDocument(<Goljus shape="5,5"/>);
     let oldBoard = goljus.getBoard();
     oldBoard[1][1] = true;
     oldBoard[1][2] = true;
@@ -189,31 +188,31 @@ it('has an advancing control', () => {
 
 it('verifies shape argument', () => {
     expect(() => {
-        TestUtils.renderIntoDocument(<Goljus shape="a"/>);
+        ReactTestUtils.renderIntoDocument(<Goljus shape="a"/>);
     }).toThrow(/correct format/);
 
     expect(() => {
-        TestUtils.renderIntoDocument(<Goljus shape="5,a"/>);
+        ReactTestUtils.renderIntoDocument(<Goljus shape="5,a"/>);
     }).toThrow(/correct format/);
 
     expect(() => {
-        TestUtils.renderIntoDocument(<Goljus shape="5"/>);
+        ReactTestUtils.renderIntoDocument(<Goljus shape="5"/>);
     }).toThrow(/correct format/);
 });
 
 it('verifies shape to be large enough', () => {
     expect(() => {
-        TestUtils.renderIntoDocument(<Goljus shape="1,3"/>);
+        ReactTestUtils.renderIntoDocument(<Goljus shape="1,3"/>);
     }).toThrow(/width needs to be at least 2/);
 
     expect(() => {
-        TestUtils.renderIntoDocument(<Goljus shape="3,1"/>);
+        ReactTestUtils.renderIntoDocument(<Goljus shape="3,1"/>);
     }).toThrow(/height needs to be at least 2/);
 });
 
 it('allows setting an update period', () => {
 
-    let goljus = TestUtils.renderIntoDocument(<Goljus shape="5,5" period={500}/>);
+    let goljus = ReactTestUtils.renderIntoDocument(<Goljus shape="5,5" period={500}/>);
     let oldBoard = goljus.getBoard();
     oldBoard[1][1] = true;
     oldBoard[1][2] = true;
@@ -243,13 +242,13 @@ it('allows setting seed', () => {
     board[1][1] = true;
     board[1][2] = true;
     board[1][3] = true;
-    let goljus = TestUtils.renderIntoDocument(<Goljus shape="5,5" seed={board} />);
+    let goljus = ReactTestUtils.renderIntoDocument(<Goljus shape="5,5" seed={board} />);
     let newBoard = goljus.getBoard();
     expect(newBoard).toBeSimilarWithBoard(board);
 });
 
 it('allows setting random seed', () => {
-    let goljus = TestUtils.renderIntoDocument(<Goljus shape="5,5" seed="random" />);
+    let goljus = ReactTestUtils.renderIntoDocument(<Goljus shape="5,5" seed="random" />);
     let newBoard = goljus.getBoard();
     expect(newBoard.height).toEqual(5);
 });
@@ -257,7 +256,7 @@ it('allows setting random seed', () => {
 it('allows setting random seed with a custom random value', () => {
     let spy = jest.fn(Goljus.randomInitializer);
     Goljus.randomInitializer = spy;
-    let goljus = TestUtils.renderIntoDocument(<Goljus shape="5,5" seed="random:30%" />);
+    let goljus = ReactTestUtils.renderIntoDocument(<Goljus shape="5,5" seed="random:30%" />);
     expect(spy).toBeCalledWith(expect.any(Number), expect.any(Number), 0.3);
     let newBoard = goljus.getBoard();
     expect(newBoard.height).toEqual(5);
@@ -265,7 +264,7 @@ it('allows setting random seed with a custom random value', () => {
 
 it('allows setting custom seed initializer', () => {
     const init = jest.fn((ii, jj) => true );
-    let goljus = TestUtils.renderIntoDocument(<Goljus shape="5,5" seed={init} />);
+    let goljus = ReactTestUtils.renderIntoDocument(<Goljus shape="5,5" seed={init} />);
     let expected = Goljus.createBoard(5, 5);
     for (let ii = 0; ii < expected.width; ii++) {
         for (let jj = 0; jj < expected.height; jj++) {
@@ -279,12 +278,12 @@ it('allows setting custom seed initializer', () => {
 });
 
 it('verifies seed argument', () => {
-    expect(() => { TestUtils.renderIntoDocument(<Goljus seed="invalid"/>);})
+    expect(() => { ReactTestUtils.renderIntoDocument(<Goljus seed="invalid"/>);})
         .toThrow(/invalid seed/);
 });
 
 it('renders a table', () => {
-    let goljus = TestUtils.renderIntoDocument(<Goljus shape="5,5" seed="random" />);
-    let table = TestUtils.findRenderedDOMComponentWithTag(goljus, 'table');
+    let goljus = ReactTestUtils.renderIntoDocument(<Goljus shape="5,5" seed="random" />);
+    let table = ReactTestUtils.findRenderedDOMComponentWithTag(goljus, 'table');
     expect(table.querySelectorAll('td').length).toEqual(25);
 });
